+++ Pluto/Classes/Core/Pluto+Login.swift
@@ -174,80 +174,4 @@ extension Pluto {
         }
     }
     
-    public func binding(type: LoginType, authString: String, success: (() -> Void)? = nil, error: ErrorCompletion? = nil) {
-        var parameters = [
-            "type": type.identifier
-        ]
-        switch type {
-        case .mail:
-            parameters["mail"] = authString
-        case .apple:
-            parameters["code"] = authString
-        case .google:
-            parameters["id_token"] = authString
-        }
-        let requestUrl = url(from: "/v1/user/binding")
-        getHeaders {
-            AF.request(
-                requestUrl,
-                method: .post,
-                parameters: parameters,
-                encoding: JSONEncoding.default,
-                headers: $0
-            ).responseJSON {
-                let response = PlutoResponse($0)
-                if response.statusOK() {
-                    success?()
-                } else {
-                    error?(response.errorCode())
-                }
-            }
-        }
-    }
-    
-    public func unbinding(type: LoginType, success: (() -> Void)? = nil, error: ErrorCompletion? = nil) {
-        guard let bindings = avialiableBindings else {
-            error?(PlutoError.notSignin)
-            return
-        }
-        guard bindings.count > 1 else {
-            error?(PlutoError.unbundNotAllow)
-            return
-        }
-        let requestUrl = url(from: "/v1/user/unbinding")
-        getHeaders {
-            AF.request(
-                requestUrl,
-                method: .post,
-                parameters: [
-                    "type": type.identifier
-                ],
-                encoding: JSONEncoding.default,
-                headers: $0
-            ).responseJSON {
-                let response = PlutoResponse($0)
-                if response.statusOK() {
-                    success?()
-                } else {
-                    error?(response.errorCode())
-                }
-            }
-        }
-    }
-    
-    public var avialiableLoginTypes: [Pluto.LoginType] {
-        var types = [Pluto.LoginType.mail, .google]
-        if #available(iOS 13.0, *) {
-            types.append(.apple)
-        }
-        // TODO: Wechat login should be removed if wechat is not installed.
-        
-        return types
-    }
-    
-    public var avialiableBindings: [PlutoUser.Binding]? {
-        DefaultsManager.shared.user?.bindings.filter {
-            avialiableLoginTypes.contains($0.loginType)
-        }
-    }
 }
