+++ PlutoSDK.podspec
@@ -8,7 +8,7 @@
 
 Pod::Spec.new do |s|
   s.name             = 'PlutoSDK'
-  s.version          = '0.4.3'
+  s.version          = '0.5'
   s.summary          = 'Swift SDK for Pluto login microservice.'
 
 # This description is used to generate tags and improve search results.
